+++ src/renderer/src/i18n/en-US/chat.json
@@ -282,7 +282,8 @@
   "mode": {
     "current": "Current mode: {mode}",
     "agent": "Agent",
-    "acpAgent": "ACP Agent"
+    "acpAgent": "ACP Agent",
+    "saveFailed": "Failed to save mode selection"
   },
   "topbar": {
     "backToParent": "Back to Parent",
