+++ src/renderer/src/i18n/es-ES/chat.json
@@ -197,7 +197,8 @@
   "mode": {
     "current": "Modo actual: {mode}",
     "agent": "Agent",
-    "acpAgent": "ACP Agent"
+    "acpAgent": "ACP Agent",
+    "saveFailed": "No se pudo guardar la selección de modo"
   },
   "topbar": {
     "backToParent": "Volver a la sesión principal",
