+++ src/renderer/src/i18n/ru-RU/chat.json
@@ -180,7 +180,8 @@
   "mode": {
     "acpAgent": "Агент АШП",
     "agent": "Агент",
-    "current": "Текущий режим: {mode}"
+    "current": "Текущий режим: {mode}",
+    "saveFailed": "Не удалось сохранить выбранный режим"
   },
   "topbar": {
     "share": "Поделиться",
