+++ src/renderer/src/i18n/tr-TR/chat.json
@@ -197,7 +197,8 @@
   "mode": {
     "current": "Geçerli mod: {mode}",
     "agent": "Agent",
-    "acpAgent": "ACP Agent"
+    "acpAgent": "ACP Agent",
+    "saveFailed": "Mod seçimi kaydedilemedi"
   },
   "topbar": {
     "backToParent": "Ebeveyne Geri Dön",
